+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release Notes for Elasticsearch Plugin
 
+## 2.0.0 - 2022-06-15
+
+### Added
+
+- Craft CMS 4 compatibility
+
+### Changed
+
+- Requires Craft CMS >= 4.0
+
 ## 1.1.0 - 2022-04-27
 
 ### Added